"None" tuple factory x 2.
This commit is the last in a commit chain defining the
"beartype.cave.NoneTypeOr" factory, a dictionary efficiently generating
and caching tuples of types containing the type of the "None" singleton
from arbitrary user-specified types and tuples of types. In theory, this
is the last commit chain for our first stable release cycle (i.e., the
much-vaunted "beartype 0.1.0"). (Delicately delicious delicatessen delicacies!)