Beartypistry x 3.
This is the real (honest) last commit in a commit chain defining the
**beartypistry** (i.e., private singleton dictionary internally
leveraged by the `@beartype` decorator to both simplify and optimize
type lookup in autogenerated wrapper functions). Specifically, this
commit generalizes the beartypistry to accept both types *and* tuples of
types. (A noncommittal pittance pits utterances minute and rancid!)