Shallow PEP 484 x 8.
This commit is the next in a commit chain adding provisional shallow
support for PEP 484 (i.e., "Type Hints") and thus the "typing" module.
Specifically, this commit augments the @callable_cached decorator widely
employed throughout the codebase to memoized arbitrary callables with
support for memoizing exceptions -- dramatically improving the safety
and applicability of that decorator. (Inept aptitudes titillate the fluidic folly!)