PEP 484 unions x 4.
This commit is the last in a commit chain fully type-checking
"typing.Union" objects. This should prove to be the most significant
commit in this repository, as this commit finalizes both deep support
for unions and shallow support for PEP-compliant type hints providing an
"__origin__" attribute whose value is a non-"typing" class, which is
most of them. (Ostentatious deliciousness!)