Shallow PEP 484 x 16.
This commit is the last in a commit chain adding provisional shallow
support for PEP 484 (i.e., "Type Hints") and thus the "typing" module.
We did it, folks. Or rather... I did it. I mean, I'd love to
self-sacrificially share the glory here, but this was all-too-sadly a
one-man volunteer show of sadistic masochism for which I can only
express profound relief that the endless litany of PEP woes and feverish
"typing" lucid dreams has nearly come full tilt to its natural terminus.
Sixteen is this number of magically delicious lucky charms after all.
(Expulsive emissions of a choroidal didacticism did act it actinically!)