Explanatory PEP exceptions x 6.
This commit is the next in a commit chain raising human-readable
exceptions when passed parameters or return values violate PEP-compliant
type hints. Specifically, this commit exhaustively tests raised
exception messages against iterables of arbitrary regular expressions
expected to both match and *NOT* match these messages. Unrelatedly, this
commit also resolves a significant edge case in the beartypistry
singleton concerning registration of the type of the "None" singleton,
which inappropriately advertises itself as a builtin when it clearly
isn't. This is why I get migraines, people. (Judicious indecision!)