Python 3.9 x 1.
This commit is the first in a commit chain adding support for Python
3.9, which naturally breaks backward compatibility in the public API of
the "typing" module in various respects. Of course it does, right? I
mean, why stabilize an API leveraged across hundreds of thousands of
open-source (to say nothing of proprietary) projects world-wide,
amiright? Specifically, this commit adds Python 3.9 to all testing
matrices, updates the `beartype.meta.PYTHON_VERSION_MINOR_MAX` integer
global to 9, and begins generalizing the codebase to handle the
inevitable idiosyncrasies introduced with Python 3.9. (Curmudgeonly cudgels!)