PEP 585 x 2.
This commit is the next in a commit chain adding support for PEP 585
(i.e., "Type Hinting Generics In Standard Collections"). Specifically,
this commit sanitizes the prior commit to comply with tedious details of
Python 3.6's non-standard "typing" module implementation as well as
narrowing the set of classes permitted to be used as signs uniquely
identifying PEP-compliant type hints for safety. (*Extracted tracts!*)