Sphinx x 1.
This commit is the first in a commit chain adding support for
Sphinx-generated documentation formatted as reStructeredText (reST) and
typically converted to HTML to be hosted at Read The Docs (RTD).
Specifically, this commit adds an initial draft Sphinx configuration
enabling the standard retinue of builtin Sphinx extensions (e.g.,
"autodoc", "viewcode") as well as MathJax configured for remote implicit
downloads and Napolean configured for NumPy-formatted docstrings.
Unrelatedly, this commit also updates all copyrights littered throughout
the codebase to reference the new year as well as adding a new prefacing
"tl;dr" section to our soon-to-be-demolished "README.rst" file. (*Tasered razors!*)