Read The Docs (RTD) badge.
This commit adds a badge (i.e., shield, status icon) to our front-facing
documentation signifying the success of the most recent attempt to build
and host our Sphinx-based documentation at Read The Docs (RTD).
Relatedly, this commit also resolves a trivial reST error with respect
to the index of that documentation. (*Ignominious gnomes of a cognoscenti's cognomen!*)