Least Recently Used (LRU) cache x 1.
This commit is the first in a commit chain implementing a highly
microoptimized Least Recently Used (LRU) cache for subsequent use
throughout the codebase, particularly with respect to caching iterators
over dictionaries, sets, and other non-sequence containers. In theory,
this commit is the last of the **beartype 0.5.2** release cycle. Yeah!
(*Situationally exciting exercise!*)