refactor: Implement threading.Lock
The **only** reason this commit is marked as refactor is:
- Change from 'RLock' to 'Lock'
- Invert logical check in `__contains__`

Everything else is geared around putting a shine on the bear's coat by
tuning the formatting of comments and doc strings. Which, in that respect,
the latter item would fall under this category as well.

Arguably the smallest change was threading.Lock. Removing the R from
'RLock' seems suspiciously easy... However, as per the FIXME, this
appears to be how `utilcacheall` implemented thread safety. Likewise
several lengthy articles seem to verify this is the case so here we are.

Inverting the logical check on '__contains__' was primarily for
readability.

Signed-off-by: Heliotrop3 <43817374+Heliotrop3@users.noreply.github.com>