Private hint data sanitization x 2.
This commit is the last in a commit chain sanitizing the private
"beartype._util.hint.data" subpackage for improved robustness. The prior
design of this subpackage was deeply flawed, promoting fragile and
subtle issues throughout the codebase that were dependent on importation
order and thus non-trivial to reproduce. Unrelatedly, this commit also
resolves a significant issue in a `mypy`-based functional test
validating `beartype` to be PEP 561-compliant. (*Buzzsaw buzzkill!*)