Data validation x 10.
This commit is the next in a commit chain adding support for arbitrary
caller-defined data validation en-route to resolving issue #32, enabling
callers to validate the internal structure of arbitrarily complex
scalars, data structures, and third-party objects. Specifically, it's so
late and it's raining and I no longer have a clear grip on what exactly
was done here but I'm fairly sure it was impressive.

## Features Optimized

* **Non-builtin types.** `@beartype` now checks non-builtin types
  optimally by avoiding an extraneous dictionary lookup in the
  beartypistry singleton previously required to check those types.

(*Outlandish outlanders!*)