Data validation x 13.
This commit is the maybe second-to-last in a commit chain adding support
for arbitrary caller-defined data validation en-route to resolving the
issue that is #32, enabling callers to validate the internal structure
of arbitrarily complex scalars, data structures, and third-party
objects. Specifically, this commit finalizes the implementation of both
the core "beartype.vale.Is" class *and* code generated to validate type
hints annotated by subscriptions of that class in `@beartype`-decorated
callables as well as unit tests exercising these facilities. The next
commit will finalize unit tests exercising all edge cases and ideally be
the last commit in this commit chain – finalizing the world's first
PEP-compliant data validation API. (*Almost almighty!*)