"beartype.vale.IsEqual" x 1.
This commit is the first in a commit chain implementing support for
stack-free validation of object equality via the newly declared
`beartype.vale.IsEqual` type hint. Specifically, this commit drafts a
preliminary implementation for that hint that has yet to be tested...
*at all.* This saddens us, too. Unrelatedly, this commit substantially
optimizes generation of the `func_wrapper_locals` dictionary of locally
scoped variables conditionally required by the current wrapper function.
(*Interminable termination!*)