"beartype.vale.IsEqual" x 2.
This commit is the last in a commit chain implementing support for
stack-free validation of object equality via the newly declared
`beartype.vale.IsEqual` type hint. Specifically, this commit both
finalizes the implementation for that hint and exhaustively tests that
hint, as well as resolving a subtle issue with dictionary collision
detection and unhashable values. (*Eruptive disruption!*)