"beartype.vale.IsAttr" x 3.
This commit is the last in a commit chain implementing support for
frame-explicit validation of object attributes via the newly declared
`beartype.vale.IsAttr` type hint. Specifically, this commit both
finalizes the implementation of that hint *and* exhaustively unit tests
that hint. In theory, this commit is the last substantive commit of the
beartype 0.7.0 release cycle. (*Outlandish misery satirizing upstanding citizenry!*)