"typing_extensions.Annotated" x 13.
This commit is the next in a commit chain adding transparent support
for the third-party `typing_extensions.Annotated` type hint back-ported
to Python < 3.9, en route to resolving #34. Once finalized, this will
enable usage of beartype validators under Python < 3.9 via this hint.
Specifically, this commit concludes all prerequisites needed to fully
implement this refactoring while still passing tests. All subsequent
commits, which are the bulk of them, will temporarily break tests as an
unavoidable side effect of the severity of this refactoring. (*Clueless clavicles!*)