Python 3.10 "NewType" x 3.
This commit is the last in a commit chain improving compatibility with
the variant of the `typing.NewType` attribute implemented under Python ≥
3.10, which fundamentally refactored that attribute from a function
factory to a class -- significantly complicating cross-version detection
of that attribute. Specifically, this commit finalizes both the
implementation and testing of this attribute under Python ≥ 3.10.
(*Ignominious gnomish minions!*)