(typing.Type|type) x 6.
This commit is the next in a commit chain adding support for deeply
type-checking type hints created by subscripting both the PEP
484-compliant `typing.Type` attribute *and* PEP 585-compliant `type`
type with arbitrary classes. Specifically, this commit finalizes our
draft implementation of this support – which has yet to be exhaustively
tested and should assumed to be suspect. (*Thaumaturgic urges!*)