Ensure README.rst is decoded as UTF-8 in test_doc_readme on all platforms
The default platform encoding on Windows is the single-byte encoding cp1252.
This causes an error when reading README.rst in test_doc_readme:

"UnicodeDecodeError: 'charmap' codec can't decode byte 0x9d in position 1495:
character maps to <undefined>"

That is the 3rd byte in the unicode LEFT DOUBLE QUOTATION MARK character from
the first code sample in the readme.