`typing.{Binary,Text,}IO[...]` x 4.
This commit is the next in a commit chain adding support for
subscripted `typing.{Binary,Text,}IO[...]` type hints, en-route to
resolving issue #75 kindly submitted by Niklas "If I had a nickel for
every lass..." Rosenstein. Specifically, this commit fully declares a
new public `beartype.vale.IsInstance` validator encapsulating the
low-level `isinstance()` builtin as a validator. Differentiating the
`typing.BinaryIO` subclass from its `typing.IO` superclass requires this
validator, as the API of the former is identical to the API of the
latter from the low-level perspective of the `isinstance()` builtin.
Naturally, this validator has yet to be either documented or tested.
(*Rampant lambency!*)