Initial attempt a caching protocol implementation
This is an initial stab at migrating
[``numerary.types.CachingProtocolMeta``](https://github.com/posita/numerary/blob/dd57bffff32c8e1c4d46409c9302318bdcfe3e20/numerary/types.py#L129)
to `beartype`. It is properly considered a work-in-progress with several open questions.
Further, as of this commit, it only implements a naive cache that will grow
indefinitely. (I.e., it does not include an LRU mechanism.)

I think it's a worthwhile starting point, though.