`@beartype` + `type` x 1.
This commit is the first in a commit chain generalizing the `@beartype`
decorator to additionally decorate classes, en-route to resolving
feature request #152 kindly submitted by @posita the positively sublime.
Specifically, this commit finalizes a draft implementation of this
generalization. Since this was surprisingly trivial (given prior commit
chains generalizing `@beartype` to additionally decorate class, static,
and property method descriptors), we probably should have done this a
few years ago. We didn't. This is why we lament into our oatmeal in
2022. Naturally, nothing was tested. Cue lamentations of the wicked!
(*Expressive compression of massive missives!*)