`beartype.abby.is_bearable()` optimization x 3.
This commit is the next in a commit chain heavily optimizing the
`beartype.abby.is_bearable()` tester. Currently, that tester behaves
reasonably optimally when the passed object satisfies the passed type
hint but *extremely* suboptimally when that object violates that hint;
this is due to our current naive implementation of that tester using the
standard Easier to Ask for Permission than Forgiveness (EAFP) approach.
Specifically, this commit is yet another biggie. Which is the opposite
of specificity, isn't it? But no one is supposed to read these commit
messages, anyway. What are you doing? Are you still reading this? You
are, aren't you!? Very well. Let's begin. Just kidding. I'm not doin'
nuthin' tonight. This is *usually* the place where I'd enumerate the
changelist for this commit with stentorian bravado. But... it's late and
I am tired. Already commit this, GitHub! (*Brave lathe!*)