`beartype.abby.is_bearable()` optimization x 4.
This commit is the next in a commit chain heavily optimizing the
`beartype.abby.is_bearable()` tester. Currently, that tester behaves
reasonably optimally when the passed object satisfies the passed type
hint but *extremely* suboptimally when that object violates that hint;
this is due to our current naive implementation of that tester using the
standard Easier to Ask for Permission than Forgiveness (EAFP) approach.
Endless journey continues endlessly. (*Rabid rapidity!*)