`beartype.door` publication x 1.
This commit is the first in a commit chain preparing the `beartype.door`
subpackage for imminent publication to the world. Specifically, this
commit publicizes *ALL* previously private `beartype.door.TypeHint`
subclasses corresponding to public PEP standards. Doing so enables end
users to introspect the classes of low-level type hints and is thus an
invaluable feature. These subclasses include:

* `beartype.door.UnionTypeHint`.
* `beartype.door.LiteralTypeHint`.
* `beartype.door.AnnotatedTypeHint`.
* `beartype.door.NewTypeTypeHint`.
* `beartype.door.TypeVarTypeHint`.
* `beartype.door.CallableTypeHint`.
* `beartype.door.TupleTypeHint`.

(*Provisional proviso is notional!*)