`beartype.door` publication x 4.
This commit is the next in a commit chain preparing the `beartype.door`
subpackage for imminent publication to the world. Specifically, this
commit augments `beartype.door.TypeHint` instances with sane `__len__()`
and `__bool__()` dunder methods enabling type hint wrappers to be
trivially sized and evaluated as booleans according to the child type
hints subscripting the wrapped type hints. Doing so enabled this commit
to also remove the questionable
`beartype.door._TupleTypeHint.is_empty_tuple` property, which is now
obsoleted in favour of simply calling ``len(type_hint)``. (*Urgency in agency!*)