`typing.NamedTuple` x 2.
This commit is the last in a commit chain adding support for deeply
type-checking subclasses of the PEP 484-compliant `typing.NamedTuple`
superclass. Specifically, this commit documents and tests this support
as well as dutifully resolving a critical edge case with Python 3.7's
non-standard implementation of `typing.NamedTuple`.
(*Essential victuals of ritual presence!*)