`BeartypeCallHintViolation.culprit` x 12.
This commit is the last in a commit chain generalizing **type-checking
violations** (i.e., instances of the
`beartype.roar.BeartypeCallHintViolation` exception class) to safely
expose the **culprits** (i.e., user-defined objects responsible for that
violation) via a new `BeartypeCallHintViolation.culprits` property,
resolving feature request #180 kindly submitted by @Jasha10 the
Supremely Patient and Understanding GitHubber. For safety, this property
dynamically returns a tuple of one or more items that are either:

* For each underlying culprit that supports weak references and is still
  alive (i.e., has yet to be garbage-collected), that culprit as is.
* Else, the machine-readable string representation of that culprit
  truncated to a reasonable number of characters.

Specifically, this commit finalizes unit tests of
`BeartypeCallHintViolation.culprits` property, which astonishingly
appears to behave as expected. (*A post-haste riposte ripe with taste!*)