`__beartype_hint__()` x 4.
This commit is the next in a commit chain defining the PEP
544-compliant `beartype.plug.BeartypeHintable` protocol, en-route to
resolving feature request #192 technically submitted by myself but
spiritually submitted by super-nice NIST scientist @tbsexton (Thurston
Sexton) in his unquenchable search for a utopian DSL for typing Python
that may never exist – *but should*. Specifically, this commit continues
iterating on the implementation, usage, and testing of this protocol.
Sadly, we suddenly realized in a fit of mad panic that this protocol
should *not* actually be a protocol. Why? Because doing so coerces *all*
user-defined classes subclassing this protocol into typing generics --
which is absolutely *not* intended. Consequently, the next commit in
this commit chain will:

* Relax `BeartypeHintable` into a standard abstract base class (ABC).
* Create a new private `_BeartypeHintableProtocol` for internal usage,
  mostly in detecting classes implicitly satisfying this ABC.

(*Thunderous terror or tons of nuns on tundra?*)