Update README.rst to specify typing imports with beartype.typing
This should probably be done first as an issue, but since the question ask is a documentation change, I've included it in the PR.
Specifically, the change also specifies the problem: if your python-fu is a wee bit lacking, you might've been doing
```
from beartype import typing
from typing import ...
``` 
All this time, and it just so happened to work, until it didn't. 
If there's a better place for this, or not at all, this commit / PR can just be deleted.