`README.rst` -> Read the Docs (RTD) x 21.
This commit is the next in a commit chain coercing our monolithic
`README.rst` documentation onto Read the Docs (RTD), en-route to
resolving issue #203 kindly submitted by @LittleBigGene (AKA the dynamo
of the cell). Specifically, this commit isolates our manually generated
API (...lol) into a new `api.rst` document *and* continues transitioning
API documentation from `README.rst` into that document. It is a heavy
weight we bear, but we bear it so that you need not. (*Lonesome burden, unordained denizen!*)