`README.rst` -> Read the Docs (RTD) x 22.
This commit is the next in a commit chain coercing our monolithic
`README.rst` documentation onto Read the Docs (RTD), en-route to
resolving issue #203 kindly submitted by @LittleBigGene (AKA the dynamo
of the cell). Specifically, this commit continues transitioning API
documentation from `README.rst` into our new `api.rst` document. It is a
heavy weight we bear, but we bear it so that you need not.
(*Zen-like netizen frolics in fields of irksome nettles!*)