`README.rst` -> Read the Docs (RTD) x 24.
This commit is the next in a commit chain coercing our monolithic
`README.rst` documentation onto Read the Docs (RTD), en-route to
resolving issue #203 kindly submitted by @LittleBigGene (AKA the dynamo
of the cell). Specifically, this commit transitions our timings and math
discussion from `README.rst` into a new `math.rst` document. It is a
heavy weight we bear, but we bear it so that you need not. (*Unflagging flagon!*)