`README.rst` -> Read the Docs (RTD) x 28.
This commit is the next in a commit chain coercing our monolithic
`README.rst` documentation onto Read the Docs (RTD), en-route to
resolving issue #203 kindly submitted by @LittleBigGene (AKA the dynamo
of the cell). Specifically, this commit temporarily disables the classic
`_templates/sidebar-nav-bs.html` template -- which sadly appears to be
silently broken on ReadTheDocs (RTD) but *not* locally. Thus, we sigh.
(*Inspirational navigation of notional spires!*)