`README.rst` -> Read the Docs (RTD) x 29.
This commit is the next in a commit chain coercing our monolithic
`README.rst` documentation onto Read the Docs (RTD), en-route to
resolving issue #203 kindly submitted by @LittleBigGene (AKA the dynamo
of the cell). Specifically, this commit unpins `pydata-sphinx-theme` to
update to arbitrarily new versions on RTD. In theory, doing so *should*
resolve the disastrously empty sidebar. In practice, doing so *will*
reduce to a noop causing us to pull even more of our hair out.
(*Dance of happy happenstance!*)