`README.rst` -> Read the Docs (RTD) x 32.
This commit is the next in a commit chain coercing our monolithic
`README.rst` documentation onto Read the Docs (RTD), en-route to
resolving issue #203 kindly submitted by @LittleBigGene (AKA the dynamo
of the cell). Specifically, this commit circumvents the dreaded empty
sidebar issue (which actually appears to be a remote RTD-specific issue
at this belated point, which is pretty much the worst-case outcome) by
simply flattening our `doc/src/` subdirectory structure (which is
insane, but welcome to 2023). Absolutely *not* happy about any of this,
but it's unclear whether anyone living is even to blame. I blame only
myself for not embracing `mkdocs` when I had the opportunity.
(*Absolved ablutions!*)