PEP 673 x 4.
This commit is the next in a commit chain adding deep type-checking
support for **PEP 673 – Self Type** (i.e., `typing.Self`). Specifically,
this commit does many, many things. We can see this, because this is a
big changeset that frightens the new guy. Because @leycec is too tired
to document things on Friday evenings, we'll just have to accept the
feeble scraps that he's throwing us here. (*Unseemly seams seem unsure!*)