`test_sphinx_docs_other()` x 1.
This commit is the first in a commit chain attempting to resolving a
recent spate of continuous integration (CI) failures pertaining to our
previously passing `test_sphinx_docs_other()` integration test, which
exercises whether @beartype integration with Sphinx's built-in `autodoc`
extension behaves as expected. Since @leycec *cannot* replicate this
issue locally, this commit chain is likely to be longer, more boring,
and keyboard-smashing than expected. Specifically, this commit
explicitly passes the `-T` command-line option to *all* invocations of
the `sphinx-build` command distributed throughout the codebase.
(*To juggle a conjugal conjugation!*)