Beartype 0.16.0 pre-flight inspection.
This commit is the last commit in our upcoming beartype 0.16.0 release
cycle. Strap in, folks. We are cleared for launch. **Beartype 0.16.0,**
better known as "Super Unsexy Stabilization Release (SUSR)," is soon to
be catapulted off the flight deck of PyPI and into the glorious
cacophony of your crowded dependency list. Prepare for incoming.
(*Bulging barges!*)