Python 3.7 last-rites x 1.
This commit is the first in a commit chain **last-riting** (i.e.,
permanently removing) support for Python 3.7, which recently hit its
official End-of-Life (EoL) date and thus constitutes a security risk.
Since Python 3.7 substantially differed from Python 3.8 with respect to
type hints in general and the standard `typing` module in specific, this
commit chain is likely to take longer than anyone wants or expects. This
is why @leycec does @beartype: so you don't have to. (*Broad roads, bro!*)