PEP 604-inconsistent type hint protection x 2.
This commit is the last in a commit chain explicitly prohibiting all
**PEP 604-inconsistent type hints** (i.e., object permissible as an item
of a PEP 604-compliant new union whose machine-readable representation
is *not* the machine-readable representation of this hint in new
unions). Specifically, this commit improves the detection of PEP
604-inconsistent type hints to be robust against unlikely edge cases
that could destroy the fabric of our fragile world like an egg popping
with a disgusting squelching noise. (*Honourable honey or an ouroboros!?*)