`beartype.BeartypeConf.hint_overrides` options x 4.
This is the next in a commit chain defining a new public
`beartype.BeartypeConf.hint_overrides` option, en-route to resolving
feature request #302 also kindly submitted by that same wizard.
Specifically, this commit is frankly trivial and best ignored. Tomorrow,
we do something meaningful. This, we swear it! (*Thus, we wear it!*)