`beartype.door.die_if_unbearable()` warning x 1.
This commit is the first in a commit chain generalizing the
`beartype.door.die_if_unbearable()` runtime type-checker to emit
non-fatal warnings when configured to do so by the passed beartype
configuration. In theory, this commit chain constitutes the last commits
of the increasingly elongated @beartype 0.17.0 release cycle.
(*Mournful horns robotically full of thrombotic thorns!*)