`beartype.door.die_if_unbearable()` warning x 3.
This commit is the next in a commit chain generalizing the
`beartype.door.die_if_unbearable()` runtime type-checker to emit
non-fatal warnings when configured to do so by the passed beartype
configuration. In theory, this commit chain constitutes the last commits
of the increasingly elongated @beartype 0.17.0 release cycle.
(*Decadent hadrons dent decadal cicadas*)