`beartype.door.die_if_unbearable()` warning x 5.
This commit is the last in a commit chain generalizing the
`beartype.door.die_if_unbearable()` runtime type-checker to emit
non-fatal warnings when configured to do so by the passed beartype
configuration. In theory, this commit chain constitutes the last commits
of the increasingly elongated @beartype 0.17.0 release cycle.
(*Flavourful flavonoids pour lava on droids!*)