`beartype.BeartypeConf.violation_type`.
*Finally.* This is it, folks. This commit constitutes the
seriously-this-is-it-the-last-commit of the @beartype 0.17.0 release
cycle, introducing a new **default violation type** (i.e., type of
exception to default whichever of the `violation_door_type`,
`violation_param_type`, and `violation_return_type` exception types are
unpassed and thus `None`). This parameter is merely a convenience
enabling users to trivially control the `violation_door_type`,
`violation_param_type`, and `violation_return_type` parameters *without*
having to explicitly pass all three of those parameters. Specifically,
if this parameter is:

* *Not* `None`, then the `violation_door_type`, `violation_param_type`,
  and `violation_return_type` parameters all default to the value of
  this parameter.
* `None`, then:
  * `violation_door_type` defaults to
    `beartype.roar.BeartypeDoorHintViolation`.
  * `violation_param_type` defaults to
    `beartype.roar.BeartypeCallHintParamViolation`.
  * `violation_return_type` defaults to
    `beartype.roar.BeartypeCallHintReturnViolation`.

Defaults to `None`. Actually, wait. `pyright` is vomitting everywhere,
so this only constitutes the
seriously-this-is-it-the-second-to-last-commit of this release cycle.
This is why we can't have sane brains. `pyright`: makin' cray-cray.
(*Based userbase!*)