Granular warning messages x 1.
This commit is the first in a commit chain improving the granularity of
warning messages, en-route to resolving feature request #73 kindly
submitted several lifetimes ago by Stuttgart `typing` master @amogorkon
(Anselm Kiefner). Specifically, this commit:

* Prefixes warning messages with contextual metadata describing the
  origin of those warnings, typically including fully-qualified module,
  class, and/or callable names as relevant.
* Associates warnings emitted by @beartype with the external
  user-defined modules to which those warnings apply under Python ≥
  3.12, whose standard `warnings.warn()` function accepts a new
  `skip_file_prefixes` parameter for exactly this purpose.

Thanks so much to @amogorkon for gently prodding us to do this several
lifetimes ago. This miracle on Earth is happening because of you.
(*Considerable sidereal dirigible!*)