Some improvements for Python 3.14 / PEP 649 support
It's still very early for Python 3.14, but I decided to try running
beartype's test suite to see how well it works with all the annotations
changes that have landed in CPython.

There were a few test failures for boring reasons (mostly, the fact
that typing.ForwardRef is now in the annotationlib module). I fixed
some but not all of those.

I also tried to make beartype work well with code that relies on
new PEP 649 functionality, such as:

from beartype import beartype

@beartype
def f(x: notyetdefined) -> notyetdefined:
    return x

class notyetdefined:
    pass

f(notyetdefined())
f(1)  # error

To support such code, I made beartype use annotationlib.get_annotations()
with the FORWARDREF format, which appears to work fine.

I will leave this as a draft for now since it's still early in the PEP 649
stage, but I feel this is a promising indicator that it shouldn't be too
hard to adapt beartype to the PEP 649 world.