Code generation objectification x 2.
This commit is the next in a commit chain fundamentally refactoring our
code generation algorithm from a raw parameter-based approach to a fully
object-oriented approach, significantly streamlining the implementation
of this algorithm. Although this commit chain is entirely of internal
interest to @beartype maintainer (...so, just @leycec then), the
maintainability gains should be significant. Let's do this, QA Bear!
(*Significant signifier!*)