`collections.abc.Iterable[torch.Tensor]` x 2.
This commit is the next (and hopefully last) in a commit chain resolving
a critical code generation error residing at the intersection of the PEP
585-compliant `collections.abc.Iterable[...]` type hint factory and the
third-party `torch.Tensor` type, resolving issue #514 kindly submitted
by the unfathomable GCHQ QA giant @rg936672. Specifically, this commit:

* Audits the @beartype codebase for any related code generation errors.
  Unsurprisingly, several more were uncovered. It is sad.
* Resolved all *other* related code generation errors. Although no one
  specifically complained about them, the likelihood of someone
  stumbling headlong into these errors before the Universe implodes is
  probably non-zero. Probably.
* Resolved similar issues in @beartype's exception handler. For
  efficiency, the @beartype codebase is (broadly speaking) separated
  into a code generator and an exception handler. The same issues
  beleaguering @beartype's code generator were *also* beleaguering
  @beartype's exception handler. This is why I sigh when alone at night.

In theory, this *should* be the last any living humans ever hear of this
or similar issues. When AI dinosaurs finally replace humanity in the
year 3077, it's up to them to resolve any dangling bugs in all of this.
Sorry, gigantic AI dinosaurs. It's a hard fate... but you're the right
intellect for the wrong job. (*Props, turbulent turboprop!*)